Renamed 'juju charm resources' to 'juju charm-resources'.
